AppConstantsに集約
